Add repository test and improve code quality

+++ src/main/scala/com/carher/payload/UserAccountResultPayload.scala
@@ -1,3 +1,3 @@
 package com.carher.payload
 
-final case class UserAccountResultPayload(count: Int)
+final case class UserAccountResultPayload(count: Long)
